Properly invite members being added to a group
If members are being added to a PMIx_Group_construct operation,
then those members need to be invited to join the group via
PMIx event. Properly track that via the grpcomm signature object,
including the info required to be passed to the invited members
so they have a complete picture of the group.
