+++ src/runtime/data_type_support/prte_dt_unpacking_fns.c
@@ -733,6 +733,43 @@ int prte_grpcomm_sig_unpack(pmix_data_buffer_t *buffer,
         return prte_pmix_convert_status(rc);
     }
 
+    // unpack the context ID, if one was assigned
+    cnt = 1;
+    rc = PMIx_Data_unpack(NULL, buffer, &s->ctxid_assigned, &cnt, PMIX_BOOL);
+    if (PMIX_SUCCESS != rc) {
+        PMIX_ERROR_LOG(rc);
+        PMIX_RELEASE(s);
+        return prte_pmix_convert_status(rc);
+    }
+    if (s->ctxid_assigned) {
+        cnt = 1;
+        rc = PMIx_Data_unpack(NULL, buffer, &s->ctxid, &cnt, PMIX_SIZE);
+        if (PMIX_SUCCESS != rc) {
+            PMIX_ERROR_LOG(rc);
+            PMIX_RELEASE(s);
+            return prte_pmix_convert_status(rc);
+        }
+    }
+
+    // unpack the added members
+    cnt = 1;
+    rc = PMIx_Data_unpack(NULL, buffer, &s->nmembers, &cnt, PMIX_SIZE);
+    if (PMIX_SUCCESS != rc) {
+        PMIX_ERROR_LOG(rc);
+        PMIX_RELEASE(s);
+        return prte_pmix_convert_status(rc);
+    }
+    if (0 < s->nmembers) {
+        PMIX_PROC_CREATE(s->addmembers, s->nmembers);
+        cnt = s->nmembers;
+        rc = PMIx_Data_unpack(NULL, buffer, s->addmembers, &cnt, PMIX_PROC);
+        if (PMIX_SUCCESS != rc) {
+            PMIX_ERROR_LOG(rc);
+            PMIX_RELEASE(s);
+            return prte_pmix_convert_status(rc);
+        }
+    }
+
     // try and unpack the groupID - error is okay, just means
     // one wasn't provided
     save = buffer->unpack_ptr;
@@ -741,7 +778,6 @@ int prte_grpcomm_sig_unpack(pmix_data_buffer_t *buffer,
     // ignore the return code
     if (PMIX_SUCCESS != rc) {
         buffer->unpack_ptr = save;  // reset location for next unpack
-        rc = PMIX_SUCCESS;
     }
 
     *sig = s;
